Initial commit. Middleware and ClassLibrary. RequestInspectors moved to Core.
