+++ nbs-manage-your-appointments.sln
@@ -11,6 +11,8 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Api", "src
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Core", "src\api\Nhs.Appointments.Core\Nhs.Appointments.Core.csproj", "{CFEAB403-3E1E-4CBE-B13D-5554029E6C18}"
 EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Nhs.Appointments.Audit", "src\api\Nhs.Appointments.Audit\Nhs.Appointments.Audit.csproj", "{D18A700C-909A-4E13-B611-E952F4806F51}"
+EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Nhs.Appointments.Persistance", "src\api\Nhs.Appointments.Persistance\Nhs.Appointments.Persistance.csproj", "{68575B3B-E68D-40AB-850F-52694C83BDD9}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "tests", "tests", "{C4AAF10B-178C-42F1-B081-22E8D5324355}"
@@ -100,6 +102,10 @@ Global
 		{B997F51F-1CDC-48FC-BC1C-C9B1706BD441}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{B997F51F-1CDC-48FC-BC1C-C9B1706BD441}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{B997F51F-1CDC-48FC-BC1C-C9B1706BD441}.Release|Any CPU.Build.0 = Release|Any CPU
+		{D18A700C-909A-4E13-B611-E952F4806F51}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{D18A700C-909A-4E13-B611-E952F4806F51}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{D18A700C-909A-4E13-B611-E952F4806F51}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{D18A700C-909A-4E13-B611-E952F4806F51}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -119,6 +125,7 @@ Global
 		{41309612-F5A4-41CF-9391-008EFAD96E80} = {22B97405-2671-42B1-BE08-E1D0E18A2F1F}
 		{00BF2F24-0FB3-420B-BE1C-830919F6EA3E} = {22B97405-2671-42B1-BE08-E1D0E18A2F1F}
 		{B997F51F-1CDC-48FC-BC1C-C9B1706BD441} = {C4AAF10B-178C-42F1-B081-22E8D5324355}
+		{D18A700C-909A-4E13-B611-E952F4806F51} = {6F9747AA-DE8B-43B4-B77F-4DDD959BB94C}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {F16CD335-B2DF-4C77-9801-5DB5B9AAFADA}
