+++ src/api/Nhs.Appointments.Api/Auth/NoSiteRequestInspector.cs
@@ -2,6 +2,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker.Http;
+using Nhs.Appointments.Core.Inspectors;
 
 namespace Nhs.Appointments.Api.Auth;
