+++ src/api/Nhs.Appointments.Api/Functions/ApplyAvailabilityTemplateFunction.cs
@@ -11,6 +11,7 @@ using Nhs.Appointments.Api.Auth;
 using Nhs.Appointments.Api.Availability;
 using Nhs.Appointments.Api.Models;
 using Nhs.Appointments.Core;
+using Nhs.Appointments.Core.Inspectors;
 
 namespace Nhs.Appointments.Api.Functions;
