+++ src/api/Nhs.Appointments.Api/Functions/GetUserPermissionsFunction.cs
@@ -11,6 +11,7 @@ using Microsoft.Azure.Functions.Worker;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.OpenApi.Models;
 using Nhs.Appointments.Api.Auth;
+using Nhs.Appointments.Core;
 
 namespace Nhs.Appointments.Api.Functions;
