+++ src/api/Nhs.Appointments.Api/Functions/GetUserRoleAssignmentsFunction.cs
@@ -13,6 +13,7 @@ using System.Net;
 using System.Threading.Tasks;
 using Microsoft.Azure.WebJobs.Extensions.OpenApi.Core.Attributes;
 using Microsoft.OpenApi.Models;
+using Nhs.Appointments.Core.Inspectors;
 
 namespace Nhs.Appointments.Api.Functions
 {
