+++ src/api/Nhs.Appointments.Api/Functions/QueryBookingsFunction.cs
@@ -10,6 +10,7 @@ using Microsoft.Extensions.Logging;
 using Nhs.Appointments.Api.Models;
 using Nhs.Appointments.Core;
 using Nhs.Appointments.Api.Auth;
+using Nhs.Appointments.Core.Inspectors;
 
 namespace Nhs.Appointments.Api.Functions;
