+++ src/api/Nhs.Appointments.Api/Functions/SetSiteAttributesFunction.cs
@@ -13,6 +13,7 @@ using Nhs.Appointments.Api.Auth;
 using Nhs.Appointments.Api.Json;
 using Nhs.Appointments.Api.Models;
 using Nhs.Appointments.Core;
+using Nhs.Appointments.Core.Inspectors;
 
 namespace Nhs.Appointments.Api.Functions;
