+++ src/api/Nhs.Appointments.Api/Nhs.Appointments.Api.csproj
@@ -46,6 +46,7 @@
 	<PackageReference Include="System.IdentityModel.Tokens.Jwt" Version="8.3.0" />
 	</ItemGroup>
   <ItemGroup>
+    <ProjectReference Include="..\Nhs.Appointments.Audit\Nhs.Appointments.Audit.csproj" />
     <ProjectReference Include="..\Nhs.Appointments.Core\Nhs.Appointments.Core.csproj" />
     <ProjectReference Include="..\Nhs.Appointments.Persistance\Nhs.Appointments.Persistance.csproj" />
   </ItemGroup>
