+++ src/api/Nhs.Appointments.Core/ClaimsExtensions.cs
@@ -1,8 +1,6 @@
-﻿using System.Collections.Generic;
-using System.Linq;
-using System.Security.Claims;
+﻿using System.Security.Claims;
 
-namespace Nhs.Appointments.Api.Auth
+namespace Nhs.Appointments.Core
 {
     public static class ClaimsExtensions
     {
