+++ src/api/Nhs.Appointments.Core/Inspectors/IRequestInspector.cs
@@ -1,9 +1,6 @@
-using System.Collections.Generic;
-using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker.Http;
 
-
-namespace Nhs.Appointments.Api.Auth;
+namespace Nhs.Appointments.Core.Inspectors;
 
 public interface IRequestInspector
 {
