+++ src/api/Nhs.Appointments.Core/Inspectors/SiteFromBodyInspector.cs
@@ -1,11 +1,7 @@
-using System.Collections.Generic;
-using System.Linq;
 using System.Text.Json;
-using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker.Http;
 
-
-namespace Nhs.Appointments.Api.Auth;
+namespace Nhs.Appointments.Core.Inspectors;
 
 public class SiteFromBodyInspector : IRequestInspector
 {
