+++ src/api/Nhs.Appointments.Core/Inspectors/SiteFromQueryStringInspector.cs
@@ -1,11 +1,6 @@
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
 using Microsoft.Azure.Functions.Worker.Http;
 
-
-namespace Nhs.Appointments.Api.Auth;
+namespace Nhs.Appointments.Core.Inspectors;
 
 public class SiteFromQueryStringInspector : IRequestInspector
 {
