+++ src/api/Nhs.Appointments.Core/Nhs.Appointments.Core.csproj
@@ -11,6 +11,7 @@
 		<PackageReference Include="Azure.Storage.Blobs" Version="12.23.0" />
 		<PackageReference Include="MassTransit" Version="8.3.4" />
 		<PackageReference Include="Microsoft.Azure.Cosmos" Version="3.45.2" />
+		<PackageReference Include="Microsoft.Azure.Functions.Worker.Core" Version="2.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Azure" Version="1.9.0" />
 		<PackageReference Include="Microsoft.Extensions.Caching.Abstractions" Version="9.0.0" />
 		<PackageReference Include="Microsoft.Extensions.Http" Version="9.0.0" />
@@ -23,6 +24,12 @@
 		<AssemblyAttribute Include="System.Runtime.CompilerServices.InternalsVisibleToAttribute">
 			<_Parameter1>Nhs.Appointments.Core.UnitTests</_Parameter1>
 		</AssemblyAttribute>
+	</ItemGroup>
+
+	<ItemGroup>
+	  <Reference Include="Microsoft.AspNetCore.Http.Abstractions">
+	    <HintPath>..\..\..\..\..\..\..\..\..\usr\local\share\dotnet\x64\shared\Microsoft.AspNetCore.App\8.0.11\Microsoft.AspNetCore.Http.Abstractions.dll</HintPath>
+	  </Reference>
 	</ItemGroup>
 
 </Project>
